refactor: consolidate generate-state-bloat into init_state module
Move GenerateStateBloat from a separate file into init_state.rs so both
state-init commands share the same module. StorageLoader and LoadStats
become private to the module.

Amp-Thread-ID: https://ampcode.com/threads/T-019d2d83-eba4-740c-8fea-04cbf1cad24f
