+++ bin/tempo/src/main.rs
@@ -35,7 +35,6 @@ static ALLOC: reth_cli_util::allocator::Allocator = reth_cli_util::allocator::ne
 static MALLOC_CONF: &[u8] = b"prof:true,prof_active:true,lg_prof_sample:19\0";
 
 mod defaults;
-mod generate_state_bloat;
 mod init_state;
 mod tempo_cmd;
